+++ workspace/linux-support/src/ioctl/_IOC_WRITE.rs
@@ -2,4 +2,4 @@
 // Copyright © 2021 The developers of linux-support. See the COPYRIGHT file in the top-level directory of this distribution and at https://raw.githubusercontent.com/lemonrock/linux-support/master/COPYRIGHT.
 
 
-const _IOC_WRITE: u64 = 1;
+const _IOC_WRITE: u32 = 1;
