+++ workspace/linux-support/src/ioctl/_IOR.rs
@@ -3,7 +3,7 @@
 
 
 /// `_IOR(a, b, c)`.
-pub const fn _IOR<c>(a: u64, b: u64) -> u64
+pub const fn _IOR<c>(a: u32, b: u32) -> i32
 {
-	_IOC(_IOC_READ, a, b, size_of::<c>() as u64)
+	_IOC(_IOC_READ, a, b, size_of::<c>())
 }
